objects models 2
work with the normal textures in blender and, despite making it worse in blender, it has improved the models in unity. I have also implemented models for, tables, chairs, shelfs, and pillars (did not fit in this update). for some reason the weapon model textures still don't seem to work. Will look into this further.