main door + boss door position fixed
All the doors where too far back, at inconsistent distances from the wall, and some had not been scaled to the proper size. These issue have been fixed. Also the boss door position was several hundred (units of measurement) of 0,0,0 in its prefab, which make positioning properly in the scene was too difficult, hence this was fixed. If this causes any issues when running the game CONTACT ME TO DISCUSS FIXING IT. Do not attempt to fix it yourself. Also there have been minor edits to the textures of the wall. Wile this may not fix the lighting issue at least all wall textures should be affected consistently. p.s if you think that the front portion of the floor of the door in front of the portcullis is clipping into the floor don't worry its supposed to look like that. It is not a mistake and if you have concerns about this feel free to contact me.